requirements.txt
openai
python-dotenv
pytest
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fastapi
uvicorn
